pyproject.toml
[project]
name = "DCCImagesPipeline"
version = "0.0.1"
dependencies = [
    "requests",
    "paramiko",
    "mysql.connector",
    'importlib-metadata; python_version>="3.9"',
]
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools", "wheel"]